--- CONTRIBUTING.md
@@ -1,2 +1,0 @@
-Please see the [clcache Wiki](https://github.com/frerich/clcache/wiki/Contributing) on guidelines
-how to contribute to the project.
