--- clcache.nuspec
@@ -1,32 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!-- Do not remove this test for UTF-8: if “O” doesn’t appear as greek uppercase omega letter enclosed in quotation marks, you should use an editor that supports UTF-8, not this one. -->
-<package xmlns="http://schemas.microsoft.com/packaging/2015/06/nuspec.xsd">
-  <metadata>
-    <id>clcache</id>
-    <title>clcache</title>
-    <version>{VERSION}</version>
-    <authors>Frerich Raabe</authors>
-    <owners>frerich,tadeu</owners> <!-- also known as package mantainers -->
-    <summary>Compiler cache for Microsoft Visual Studio</summary>
-    <description>
-A little Python script which attempts to avoid unnecessary recompilation by reusing previously cached object files if possible. It is meant to be called instead of the original 'cl.exe' executable. The script analyses the command line to decide whether source code is to be compiled. If so, a cache will be queried for a previously stored object file.
-
-This package contains an executable generated by `pyinstaller`, and is meant to facilitate the usage of `clcache.py` in a portable way. It is specially useful when needing to use `clcache.py` in environments with Python 2, since newer versions of `clcache.py` supports only Python 3.4+.
-    </description>
-    <projectUrl>https://github.com/frerich/clcache</projectUrl>
-    <packageSourceUrl>https://github.com/frerich/clcache</packageSourceUrl>
-    <projectSourceUrl>https://github.com/frerich/clcache</projectSourceUrl>
-    <tags>clcache clcache.py ccache compiler</tags>
-    <copyright>
-Copyright (c)
-  2010, 2011, 2012, 2013, 2016 froglogic GmbH
-  2016 Simon Warta (Kullo GmbH)
-  2016 Tim Blechmann
-    </copyright>
-    <licenseUrl>https://github.com/frerich/clcache/blob/master/LICENSE</licenseUrl>
-    <requireLicenseAcceptance>false</requireLicenseAcceptance>
-  </metadata>
-  <files>
-    <file src="clcache-{VERSION}\**" target="clcache" />
-  </files>  
-</package>
