--- clcache.pth
@@ -1,1 +1,0 @@
-import sys; import site; sys.path.extend(site.getsitepackages()); import clcache.monkey; clcache.monkey.main()
