+++ clcache/clcache_lib/clcache_lib/utils/errors.py
@@ -26,8 +26,14 @@ class ExternalDebugInfoError(AnalysisError):
 class CalledForPreprocessingError(AnalysisError):
     pass
 
+
 class CalledForJsonOutputError(AnalysisError):
     pass
+
+
+class CalledWithoutOutputFile(AnalysisError):
+    pass
+
 
 class InvalidArgumentError(AnalysisError):
     pass
