--- clcachesrv.py
@@ -1,7 +1,0 @@
-import sys
-import os
-
-sys.path.insert(0, os.path.dirname(__file__))
-
-from clcache.server.__main__ import main
-main()
