--- moccache/__init__.py
@@ -1,1 +1,0 @@
-from .__main__ import VERSION
