--- pyinstaller/clcache_main.py
@@ -1,3 +1,0 @@
-from clcache.__main__ import main_wrapper
-import clcache.storage 
-main_wrapper()
