--- server/requirements.txt
@@ -1,1 +1,0 @@
-pyuv>=1.3.0,<=1.4.0
